refactor(probeservices): move randomPassword in register.go
Part of https://github.com/ooni/probe-engine/issues/651
