+++ probeservices/register.go
@@ -2,6 +2,8 @@ package probeservices
 
 import (
 	"context"
+	"math/rand"
+	"time"
 )
 
 type registerRequest struct {
@@ -36,4 +38,15 @@ func (c Client) MaybeRegister(ctx context.Context, metadata Metadata) error {
 	state.ClientID = resp.ClientID
 	state.Password = pwd
 	return c.StateFile.Set(state)
+}
+
+func randomPassword(n int) string {
+	// See https://stackoverflow.com/questions/22892120
+	const letterBytes = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
+	rnd := rand.New(rand.NewSource(time.Now().UnixNano()))
+	b := make([]byte, n)
+	for i := range b {
+		b[i] = letterBytes[rnd.Intn(len(letterBytes))]
+	}
+	return string(b)
 }
